create package filter mentor
