+++ src/main/java/school/faang/user_service/filter/mentor/RequestFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.mentor;
 
 import school.faang.user_service.dto.mentor.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
